Refactor : edit toast & login & join
- 불필요한 주석 및 코드 삭제
- session.storage를 사용한 자동 로그인 관련 기능 제거
