+++ src/elements/Toast.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
@@ -13,52 +11,16 @@ const autoClose = 1000;
 export const notify = ({ type, text }: Props) => {
   switch (type) {
     case "default":
-      toast(text, {
-        position: "top-center",
-        autoClose: autoClose,
-        hideProgressBar: true,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-        theme: "light",
-      });
+      toast(text);
       break;
     case "success":
-      toast.success(text, {
-        position: "top-center",
-        autoClose: autoClose,
-        hideProgressBar: true,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-        theme: "light",
-      });
+      toast.success(text);
       break;
     case "warning":
-      toast.warning(text, {
-        position: "top-center",
-        autoClose: autoClose,
-        hideProgressBar: true,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-        theme: "light",
-      });
+      toast.warning(text);
       break;
     case "error":
-      toast.error(text, {
-        position: "top-center",
-        autoClose: autoClose,
-        hideProgressBar: true,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-        theme: "light",
-      });
+      toast.error(text);
       break;
   }
 };
