--- src/hooks/hooks.md
@@ -1,3 +1,0 @@
-커스텀 훅이 위치하는 폴더입니다.
-
-ex) react-query를 이용한 hook들이 들어가는 곳
